requirements.txt
diagrams==0.24.4
PyJWT==2.10.1
bcrypt==4.3.0
boto3==1.38.41
graphviz==0.20.1